# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/users/register":{"post":{"operationId":"UsersController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]}]}},"/api/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}},"/api/users/deleted":{"get":{"operationId":"UsersController_findOnlyDeleted","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}},"/api/users/{id}/edit":{"put":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}},"/api/users/{id}/delete":{"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}},"/api/users/{id}/approve":{"patch":{"operationId":"UsersController_approve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}},"/api/users/{id}/restore":{"patch":{"operationId":"UsersController_restore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"],"security":[{"X-API-KEY":[]},{"bearer":[]}]}}},"info":{"title":"API SERVICE WEB","description":"API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"Bearer","bearerFormat":"Token","name":"Authorization","type":"http"},"X-API-KEY":{"type":"apiKey","in":"header","name":"X-API-KEY"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirm_password":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}},"required":["name","email","password","confirm_password","roles"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"},"confirm_password":{"type":"string"}},"required":["name","password","confirm_password"]}}}}